requirements.txt
torch==2.0.1
transformers
gradio==3.35.2
datasets
librosa
ffmpeg-python
python-dotenv
aiohttp
